requirements.txt
google-generativeai==0.3.2
boto3==1.34.0
requests==2.31.0
python-multipart==0.0.6 